+++ services/index.js
@@ -1,16 +1,1 @@
-import { authFetcher } from "./authFetcher"
-import { AUTHENTICATE_MUTATION } from "./GQL"
-
-export const GQLService = {
-  authenticateUser: async ({ email }) => {
-    const { authenticate } = await authFetcher({
-      query: AUTHENTICATE_MUTATION,
-      variables: {
-        input: {
-          email,
-        },
-      },
-    })
-    return authenticate
-  },
-}
+export * from "./auth"
